fix: merge history + history_go
